requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.7.2
pandas==1.2.5
numpy==1.21.0
scipy==1.7.0
matplotlib==3.3.0
pillow>=6.2.0
PyYAML==5.4.1